+++ Sources/Extensions/Codable/Encodable+Extensions.swift
@@ -35,4 +35,20 @@ public extension Encodable {
 
         return json
     }
+
+    /// Method that converts encodable model to URLQueryItems array
+    /// - Returns: URLQueryItems array
+    func asUrlQueryItems() throws -> [URLQueryItem] {
+        return try toJSON().map {
+            if ($1 is [String: Any] || $1 is [Any]),
+                let jsonData = try? JSONSerialization.data(withJSONObject: $1, options: []),
+                let jsonString = String(data: jsonData, encoding: .utf8) {
+                return URLQueryItem(name: $0, value: jsonString)
+            } else if let value = $1 as? CustomStringConvertible {
+                return URLQueryItem(name: $0, value: value.description)
+            } else {
+                throw LeadKitError.failedToEncodeQueryItems
+            }
+        }
+    }
 }