+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 0.9.25
+- **Add**: `queryItems` parameter for `ApiRequestParameters`.
+- **Add**: `asQueryItems` method for `Encodable` that converts model to query items array.
+
 ### 0.9.24
 - **Fix**: Make `ApiRequestParameters` properties public.